Cache bootloader update to reduce chance of bricking.
This change is incomplete but the caching alone is probably worth
doing as an initial step.
Acually writing the bootloader is commented out at the moment.

I had been thinking about adding code ala hubtools ported to no_std to
do signature checking in update_server, but that would be a third
implementation of signature checking and would further bloat update_server.

Cliff suggests staging the new bootloader image in the unused
bootloader-reserved flash area.
This would allow a signature check on the staged bootloader on the next
reboot. That signature status would be made available to update_server
and MGS, allowing us to avoid bricking due to incorrectly signed images.

The flow would become:
  - prep stage_bootloader
  -  write 0..n
  - finish
  - reset rot
  - check status info for valid staged bootloader
  - if staged bootloader is valid
      - prep activate_bootloader
      - finish
          - update_server can know that staged has not been written
	    since reset
          - update_server copies staged to active bootloader
      - check integrity (MGS intended image, active image, and staged
	image should all have the same hash.)
      - reset
  - if fail
      - Active bootloader has not been touched.